+++ newIDE/app/src/stories/index.js
@@ -75,7 +75,7 @@ import InstructionSelector from '../EventsSheet/InstructionEditor/InstructionOrE
 import ParameterRenderingService from '../EventsSheet/ParameterRenderingService';
 import { ErrorFallbackComponent } from '../UI/ErrorBoundary';
 import CreateProfile from '../Profile/CreateProfile';
-import ProfileDetails from '../Profile/ProfileDetails';
+import ProfileEditor from '../Profile/ProfileEditor';
 import LimitDisplayer from '../Profile/LimitDisplayer';
 import ResourcePreview from '../ResourcesList/ResourcePreview';
 import ResourcesList from '../ResourcesList';
@@ -3935,17 +3935,17 @@ storiesOf('LimitDisplayer', module)
     />
   ));
 
-storiesOf('ProfileDetails', module)
+storiesOf('ProfileEditor', module)
   .addDecorator(paperDecorator)
   .addDecorator(muiDecorator)
   .add('profile', () => (
-    <ProfileDetails
+    <ProfileEditor
       authenticatedUser={fakeIndieAuthenticatedUser}
       onEditProfile={action('edit profile')}
     />
   ))
   .add('loading', () => (
-    <ProfileDetails
+    <ProfileEditor
       authenticatedUser={fakeAuthenticatedButLoadingAuthenticatedUser}
       onEditProfile={action('edit profile')}
     />
